backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
dj-database-url==2.1.0
django-environ==0.11.2
psycopg2-binary==2.9.9
gunicorn==22.0.0
razorpay==1.3.0
